requirements.txt
streamlit
tensorflow==2.3.0
numpy
pillow
streamlit-drawable-canvas